Added a temporary route to test the buttons.
